backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
gunicorn>=21.2,<22.0
